requirements.txt
Django==4.2.7
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
Pillow==10.0.1
reportlab==4.0.4
djangorestframework==3.14.0
drf-yasg==1.21.7
python-barcode==0.15.1
django-filter==23.5
django-tables2==2.7.0
django-widget-tweaks==1.5.0